Add files via upload
Articlelist.php: The Controller manages the user requests received as HTTP GET by calling the appropriate model for the task and then selects the proper view. In this case it calls the model defined in the file 'models/upvotehistory.php' (selects the data from the database for the list of articles a user has upvoted between 2 dates) and selects the view defined in 'views/articlelist.php (displays the data onto the page).