uploaded articlelistview.php
The view has also renamed from articlelist.php to articlelistview.php as it represents the view of the article list page. A new button has been created so that when the user clicks on it, they will get a list of articles they wrote between two dates. The table display will vary depending on which button the user clicked (upvotes or written articles).