Added tag and title filters to both votes and written articles
For both Written Articles and Voted Articles queries, the $tag and $title filters from the controller have been used to filter query results. If the user types a tag or a title or both into the form, the result will return articles matching their request.
The Written Articles query has also been re-written so that it does not rely on the "re-" prefix in the permlink to filter out comments. Rather, the TxVotes table has been joined with the Comments table using the author column as well as the permlink column, to use the depth field in the Comments table to filter out comments.