searchtag() function written for searching multiple tags in database
Added a searchtag() function for use in both written and voted article modes. It searches through the database for multiple tags by using a for loop to add conditions to SQL. It uses the NEAR function to find matching tags in json_metadata so dApps storing info in json_metadata won't interfere with tag search.